rename package to adapters
