+++ consensus/src/test/kotlin/maru/consensus/qbft/adapters/QbftValidatorProviderAdaptorTest.kt
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package maru.consensus.qbft.adaptors
+package maru.consensus.qbft.adapters
 
 import kotlin.test.Test
 import maru.consensus.ValidatorProvider
